Subtle improvement to initial basis handling
This is a very slight possibility that this will change how slots
are chosen.

The difference between this new version and the old code is that
there is a 1 in 2^32 chance that a nonce generated by way of
the placement algorithm will come up as 0; in this new code, it
will be allowed be 0 as the input to the hash function, rather
than being replaced by the initial basis value.